### 2019-08-12 986e402
a bunch of text updates
