### 2019-08-12 #1641 @danielskatz
@paul-buerkner - while I find an editor, I've also suggested a number of text changes in https://github.com/paul-buerkner/thurstonianIRT/pull/2 - if you merge this, please regenerate the pdf by entering ```@whedon generate pdf``` as a new comment in this thread.
