### 2019-08-12 #1641 @paul-buerkner
@bgoodri may be able to review this submission as he is responsible for the rstan interface to Stan on which thurstonianIRT (partially) relies on.
