### 2019-08-12 40fb83f
remove pdf of the paper
