### 2019-08-12 #1641 @paul-buerkner
@yrosseel may be able to review this submission as he is responsible for the lavaan package on which thurstonianIRT (partially) relies on.
