### 2019-08-14 #1641 @TomDonoghue
I would if I could, but I don't think I'm right for this one - I only know minimal R, and I'm not particularly familiar with the stats they implement. I'm going to have to pass on this one - I'm sure someone else could do a much better job!
