### 2019-08-16 #1641 @DominiqueMakowski
Same for me, IRT is not really within my comfort zone... will have to pass! Authors of IRT-related packages might also be interested (e.g. [1](https://github.com/xluo11/xxIRT), [2](https://github.com/philchalmers/mirt), [3](https://github.com/RobertMyles/IRT), [4](https://github.com/tminka/ItemResponse)) ☺️
