### 2019-08-19 #1662 @whedon
[REVIEW]: thurstonianIRT: Thurstonian IRT Models in R
**Submitting author:** @paul-buerkner (<a href="http://orcid.org/0000-0001-5765-8995">Paul-Christian Bürkner</a>)
**Repository:** <a href="https://github.com/paul-buerkner/thurstonianIRT" target ="_blank">https://github.com/paul-buerkner/thurstonianIRT</a>
**Version:** 0.10.0
