### 2019-08-22 #1662 @russellpierce
@tomfaulkenberry do you have a copy of Mplus that you can use to evaluate the functional claims of the authors? I do not, and have limited familiarity, so I'm prone to leave that entirely with you. Okay?
