### 2019-08-22 #1662 @paul-buerkner
In the worst case I may also remove the Mplus part from the paper if it is
hard to get a copy of it. After all it is not open source or free to use.

Russell S. Pierce <notifications@github.com> schrieb am Do., 22. Aug. 2019,
05:08:

> @tomfaulkenberry <https://github.com/tomfaulkenberry> do you have a copy
> of Mplus that you can use to evaluate the functional claims of the authors?
> I do not, and have limited familiarity, so I'm prone to leave that entirely
> with you. Okay?
>
