### 2019-08-30 #1662 @tomfaulkenberry
Sorry for my late addition here...no, I do not have Mplus either.  I do like Paul's idea of just removing the Mplus part, as it does go against the spirit of open source software.

We just started our Fall term this week, so I am a bit behind on doing this full review.  I will have it done in the coming week, though.
