### 2019-09-02 #1662 @paul-buerkner
I will remove the Mplus part from the paper in the next few days but feel free to review the rest already and just ignore the Mplus part.
