### 2019-09-03 paul-buerkner/thurstonianIRT#3 @tomfaulkenberry
fitting models in lavaan doesn't seem to work
Using the "triplets" data from the documentation, I get the following error:

```
> fit <- fit_TIRT_lavaan(triplets_long)
Warning message:
In lav_model_estimate(lavmodel = lavmodel, lavpartable = lavpartable,  :
  lavaan WARNING: the optimizer warns that a solution has NOT been found!
```

This error appears for both the CRAN version as well as the developer version.

Note that using `fit_TIRT_stan` works fine.
