### 2019-09-03 #1662 @tomfaulkenberry
I have completed my review of thurstonianIRT.  I find it to be a powerful, well-documented package that will likely be of great use to many researchers interested in IRT modeling on forced-choice questionnaires.  I am impressed by the simple interface, particularly as it calls Stan "under the hood" without the need for a beginner to get too deep into Stan coding.

Only one issue came up for me when testing the software.  I could not get the `fit_TIRT_lavaan` function to work (either for the CRAN version or the developer version).  I've logged this issue here: https://github.com/paul-buerkner/thurstonianIRT/issues/3#issue-488603450.  Please feel free to ask me more questions about setup, etc., and I'll be happy to send relevant details.

Once this issue is cleared up for me, I will be more than happy to endorse publication.

Tom
