+++ man/triplets.Rd
@@ -4,7 +4,7 @@
 \name{triplets}
 \alias{triplets}
 \title{Triplets of Pairwise Comparisons}
-\format{A data frame of 100 observations containing
+\format{A data frame of 200 observations containing
   information on 12 variables.
   Overall, the 12 items measure 3 different traits.
   Items 1, 4, 7, and 10 load on trait 1,
@@ -35,7 +35,7 @@ triplets
 }
 \description{
 This data set contains synthetic data
-of 100 participants on 4 triplets. In each triplet,
+of 200 participants on 4 triplets. In each triplet,
 participants had to rank the three alternative items according
 to their preference. Reponses were then converted into
 a set of dichotomous pairwise responses between all the
