### 2019-09-03 paul-buerkner/thurstonianIRT#3 @paul-buerkner
Thanks! The subset of the original data that I used for the example `triplets` data set turned out to be too small so that lavaan ran into convergence issues. I extended the example data from 100 to 200 participants and now lavaan seems to converge nicely. Can you confirm with the latest dev version of the package?
