### 2019-09-03 #1662 @paul-buerkner
Thanks for the comments, @tomfaulkenberry! The issue you raised should be fixed now.

With regard to Mplus, I realize that I actually don't mention it in the paper already so
there may not be any need to change. Still, thurstonianIRT itself needs to support Mplus
for comparison and validation of former implementations which were all written in Mplus.
Since I don't make any claims about it in the paper, I hope it's fine that the reviewers cannot test the Mplus implementation of thurstonianIRT themselves.
