+++ README.md
@@ -40,7 +40,7 @@ library(thurstonianIRT)
 ```
 
 As a simple example consider a data set of 4 blocks each containing 3
-items (i.e., triplets) answered by 100 participants.
+items (i.e., triplets) answered by 200 participants.
 
 ``` r
 data("triplets")
@@ -119,12 +119,12 @@ head(pr)
 #> # A tibble: 6 x 6
 #>      id trait estimate    se lower_ci upper_ci
 #>   <int> <chr>    <dbl> <dbl>    <dbl>    <dbl>
-#> 1     1 t1      0.360  0.544   -0.623    1.50 
-#> 2     1 t2     -1.49   0.567   -2.64    -0.454
-#> 3     1 t3      0.0196 0.603   -1.06     1.29 
-#> 4     2 t1     -0.868  0.536   -1.98     0.165
-#> 5     2 t2      0.654  0.583   -0.409    1.91 
-#> 6     2 t3      0.383  0.607   -0.723    1.62
+#> 1     1 t1       0.301 0.502   -0.638  1.28   
+#> 2     1 t2      -1.29  0.568   -2.44  -0.266  
+#> 3     1 t3       0.340 0.549   -0.699  1.44   
+#> 4     2 t1      -0.982 0.527   -2.05  -0.00288
+#> 5     2 t2       0.880 0.608   -0.295  2.16   
+#> 6     2 t3       0.665 0.584   -0.474  1.89
 ```
 
 The thurstonianIRT package not only comes with model fitting functions
@@ -134,7 +134,7 @@ models. Below we simulate data with a very similar structure to the
 
 ``` r
 sim_data <- sim_TIRT_data(
-  npersons = 100,
+  npersons = 200,
   ntraits = 3,
   nblocks_per_trait = 4,
   gamma = 0,
