### 2019-09-04 paul-buerkner/thurstonianIRT#3 @tomfaulkenberry
Excellent...yes, everything works as claimed now, so I am closing the issue.  I appreciate the quick work, Paul!
