### 2019-09-04 #1662 @tomfaulkenberry
Thanks Paul...I closed the issue referenced above.  All has checked out, and I am quite satisfied with everything.  My review is now complete!
