### 2019-09-13 5aac9a8
improve error message when accidentally using duplicated item names.
