### 2019-09-25 6180ff4
fix lavaan and mplus code for blocks of two items
