### 2019-09-25 c99269b
allow to fit gaussian models with mplus
