### 2019-09-30 paul-buerkner/thurstonianIRT#4 @russellpierce
Set a release tag 
> Version: Does the release version given match the GitHub release (0.9.0)?

Github doesn't have any 'release' set and the current version is 0.9.0.  I see the current version is 0.9.1.  Please consider this issue a TODO to bump version # and tag a release upon completion of the review.
