### 2019-09-30 paul-buerkner/thurstonianIRT#4 @paul-buerkner
Thank you! Will do.
