### 2019-10-04 e6e493b
DESCRIPTION: bump R version requirement #6
