### 2019-10-04 6fe03b5
update doc of 'set_block'
