+++ man/set_block.Rd
@@ -23,7 +23,11 @@ if the same item was used in different blocks.}
 }
 \description{
 Prepare blocks of items and incorporate information
-about which item belongs to which trait
+about which item belongs to which trait. A block
+of items is a set of two or more items presented and answered together
+by fully ranking them or selecting the most and/or least favorit
+in a forced choice format. A whole test usually contains
+several blocks and items may reappear in different blocks.
 }
 \examples{
 set_block(
