### 2019-10-04 0265bd3
correctly handle SEs of trait scores as returned by Mplus
