+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: thurstonianIRT
 Encoding: UTF-8
 Type: Package
 Title: Thurstonian IRT Models
-Version: 0.9.2
+Version: 0.9.3
 Date: 2019-10-04
 Authors@R: c(person("Paul-Christian", "Bürkner", 
                     email = "paul.buerkner@gmail.com",
@@ -34,7 +34,8 @@ Imports:
     mvtnorm,
     utils,
     stats,
-    rlang
+    rlang,
+    knitr
 Suggests:
     testthat (>= 0.9.1),
 LinkingTo: 
@@ -43,6 +44,8 @@ LinkingTo:
     BH (>= 1.66.0-1), 
     Rcpp (>= 0.12.16), 
     RcppEigen (>= 0.3.3.4.0)
+VignetteBuilder: 
+    knitr
 SystemRequirements: 
     GNU make
 URL: https://github.com/paul-buerkner/thurstonianIRT
