### 2019-10-04 f31b2f6
add paramter recovery vignette #8
