+++ vignettes/TIRT_sim_tests.Rmd
@@ -1,0 +1,199 @@
+---
+title: "Test parameter recovery via simulations with thurstonianIRT"
+author: "Paul Bürkner"
+date: "`r Sys.Date()`"
+output:
+  rmarkdown::html_vignette:
+  toc: yes
+vignette: >
+  %\VignetteIndexEntry{Test parameter recovery via simulations with thurstonianIRT}
+  \usepackage[utf8]{inputenc}
+  %\VignetteEngine{knitr::rmarkdown}
+params:
+  EVAL: !r identical(Sys.getenv("NOT_CRAN"), "true")
+editor_options: 
+  chunk_output_type: console
+---
+
+```{r, SETTINGS-knitr, include=FALSE}
+stopifnot(require(knitr))
+options(width = 90)
+opts_chunk$set(
+  comment = NA,
+  message = FALSE,
+  warning = FALSE,
+  eval = if (isTRUE(exists("params"))) params$EVAL else FALSE,
+  dev = "png",
+  dpi = 150,
+  fig.asp = 0.8,
+  fig.width = 5,
+  out.width = "60%",
+  fig.align = "center"
+)
+```
+
+# Introduction
+
+In this vignette, we will perform a small simulation study to
+test whether the model fitting engines implemented in the thurstonianIRT
+package are able to recover known parameter values from a simulated data set.
+This also extends the automated unit tests implemented in the package
+to check the correctness of the software. For a more extensive simulation
+study using thurstonianIRT, see Bürkner, Schulte, and Holling (2019).
+
+First, we will load a bunch of packages required in the vignette.
+```{r}
+library(thurstonianIRT)
+library(dplyr)
+library(tidyr)
+```
+
+Next, we will set up the simulation condition.
+```{r}
+set.seed(1234)
+npersons <- 500
+ntraits <- 5
+nitems_per_block <- 3
+nblocks_per_trait <- 9
+nblocks <- ntraits * nblocks_per_trait / nitems_per_block
+nitems <- ntraits * nblocks_per_trait
+ncomparisons <- (nitems_per_block * (nitems_per_block - 1)) / 2 * nblocks
+```
+
+Now, we will choose a set of true parameter values.
+```{r}
+set.seed(1234)
+lambda <- runif(nitems, 0.65, 0.96)
+signs <- c(rep(1, ceiling(nitems / 2)), rep(-1, floor(nitems / 2)))
+lambda <- lambda * signs[sample(seq_len(nitems))]
+gamma <- runif(nitems, -1, 1)
+Phi <- diag(5)
+```
+
+Finally, we put all of this together and simulate a data set based
+on the condion and true parameter values.
+```{r}
+sdata <- sim_TIRT_data(
+  npersons = npersons, 
+  ntraits = ntraits, 
+  nitems_per_block = nitems_per_block,
+  nblocks_per_trait = nblocks_per_trait,
+  gamma = gamma,
+  lambda = lambda,
+  Phi = Phi
+)
+```
+
+We know that the chosen simulation condition and parameter values are well
+behaved (this may not be the case in all situations; see Bürkner, Schulte, &
+Holling, 2019). Accordingly, the model fitting engines should show good
+parameter recovery given that they are correctly implemented. We will now go
+ahead and fit the model with all three engines.
+```{r, results="hide"}
+fit_stan <- fit_TIRT_stan(sdata, chains = 1, iter = 1000, warmup = 500)
+fit_lavaan <- fit_TIRT_lavaan(sdata)
+fit_mplus <- fit_TIRT_mplus(sdata)
+```
+
+We can now predict the trait scores of all persons and compare them
+to the true scores, which are stored in the simulated data set.
+```{r}
[…]
