### 2019-10-04 ce743f6
mention advantages of IRT in both README and paper #5
