### 2019-10-04 paul-buerkner/thurstonianIRT#9 @russellpierce
Update Requested: Community Guidelines
With regard to these checklist items...

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

2 and 3 seem partially contained within the DESCRIPTION file and will render well enough on CRAN.  However, if it isn't too much to ask, then I should like to see all three handled in the repo README as well.  I expect they'll be a bit more discoverable there and the presence of #2 in particular may encourage novices to engage and contribute to open source.
