### 2019-10-04 paul-buerkner/thurstonianIRT#9 @paul-buerkner
Agreed. Will fix that!
