### 2019-10-04 #1662 @russellpierce
Pending the completion of https://github.com/paul-buerkner/thurstonianIRT/issues/9 and https://github.com/paul-buerkner/thurstonianIRT/issues/4  I too am now quite satisfied with everything. My review is now complete and I endorse publication!
