### 2019-10-05 726d78e
README: add FAQ #9
