### 2019-10-05 paul-buerkner/thurstonianIRT#9 @paul-buerkner
I just added an FAQ section to the README answering all three aspects.
