+++ NEWS.md
@@ -1,3 +1,21 @@
+# thurstonianIRT 0.10.0
+
+## Bug Fixes
+
+* Correctly handle standard errors of trait scores as returned by Mplus.
+* Fix the `triplets` example data to work with lavaan and Mplus without
+convergence issues.
+
+## New Features
+
+* Support family `gaussian` when using lavaan as the model fitting engine.
+
+## Other Changes
+
+* Improve the documentation and presentation of the package across the board 
+thanks to Russell S. Pierce, Thomas J. Faulkenberry, and Daniel S. Katz.
+
+
 # thurstonianIRT 0.9.0
   
 * Initial release version
