### 2019-10-05 paul-buerkner/thurstonianIRT#4 @paul-buerkner
I bumped the version number to 0.10.0 and added a github release.
