### 2019-10-05 #1662 @paul-buerkner
Thank you @russellpierce for your very thorough and helpful review! I have incorporated all the requested changes and think the software and related paper are now ready.
