### 2019-10-08 a04dede
Added set_blocks_from_df function
