### 2019-10-08 a810ef0
Refined set_blocks_from_df example and minor tweaks
