### 2019-10-08 e475c73
clean up 'set_blocks_from_df'
