+++ R/misc.R
@@ -100,8 +100,19 @@ named_list <- function(names, values = NULL) {
   setNames(values, names)
 }
 
+# coerce 'x' to a single character string
+as_one_character <- function(x, allow_na = FALSE) {
+  s <- substitute(x)
+  x <- as.character(x)
+  if (length(x) != 1L || anyNA(x) && !allow_na) {
+    s <- deparse_combine(s, max_char = 100L)
+    stop2("Cannot coerce ", s, " to a single character value.")
+  }
+  x
+}
+
+# coerce 'x' to a signle number value
 as_one_numeric <- function(x, allow_na = FALSE) {
-  # coerce 'x' to a signle number value
   s <- substitute(x)
   x <- suppressWarnings(as.numeric(x))
   if (length(x) != 1L || anyNA(x) && !allow_na) {
@@ -111,8 +122,8 @@ as_one_numeric <- function(x, allow_na = FALSE) {
   x
 }
 
+# coerce 'x' to TRUE or FALSE if possible
 as_one_logical <- function(x, allow_na = FALSE) {
-  # coerce 'x' to TRUE or FALSE if possible
   s <- substitute(x)
   x <- as.logical(x)
   if (length(x) != 1L || anyNA(x) && !allow_na) {
@@ -122,8 +133,8 @@ as_one_logical <- function(x, allow_na = FALSE) {
   x
 }
 
+# combine deparse lines into one string
 deparse_combine <- function(x, max_char = 100) {
-  # combine deparse lines into one string
   out <- collapse(deparse(x))
   if (isTRUE(max_char > 0)) {
     out <- substr(out, 1, max_char)
