+++ NAMESPACE
@@ -17,6 +17,7 @@ export(make_mplus_code)
 export(make_sem_data)
 export(make_stan_data)
 export(set_block)
+export(set_blocks_from_df)
 export(sim_TIRT_data)
 import(Rcpp)
 import(dplyr)
