### 2019-10-18 #1662 @cMadan
@paul-buerkner, can you look over your bib file? It looks like there are a few mistakes to fix (missing volume/pages, incorrect page numbers, missing capitalisation).
