+++ paper/paper.bib
@@ -1,28 +1,34 @@
 @article{brown2011,
-author = {Brown, Anna  and Maydeu-Olivares, Alberto},
-title ={Item Response Modeling of Forced-Choice Questionnaires},
-journal = {Educational and Psychological Measurement},
-volume = {71},
-pages = {460-502},
-year = {2011},
-doi = {10.1177/0013164410375112},
+  author = {Brown, Anna  and Maydeu-Olivares, Alberto},
+  title = {Item Response Modeling of Forced-Choice Questionnaires},
+  journal = {Educational and Psychological Measurement},
+  volume = {71},
+  number = {3},
+  pages = {460--502},
+  year = {2011},
+  doi = {10.1177/0013164410375112},
 }
 
 @article{buerkner2019,
-  title={On the statistical and practical limitations of Thurstonian IRT models},
+  title={On the Statistical and Practical Limitations of Thurstonian {IRT} Models},
   author={B{\"u}rkner, Paul-Christian and Schulte, Niklas and Holling, Heinz},
   journal={Educational and Psychological Measurement},
-  pages={0013164419832063},
   year={2019},
+  pages = {827--854},
+  volume = {79},
+  number = {5},
   publisher={SAGE Publications Sage CA: Los Angeles, CA},
   doi = {10.1177/0013164419832063},
 }
 
 @Article{carpenter2017,
   author    = {Carpenter, B. and Gelman, A. and Hoffman, M. and Lee, D. and Goodrich, B. and Betancourt, M. and Brubaker, M. A. and Guo, J. and Li, P. and Ridell, A.},
-  title     = {Stan: A Probabilistic Programming Language},
+  title     = {{S}tan: A Probabilistic Programming Language},
   journal   = {Journal of Statistical Software},
   year      = {2017},
+  pages = {1--32},
+  volume = {76},
+  number = {1},
   owner     = {Paul},
   timestamp = {2015.06.19},
   doi = {10.18637/jss.v076.i01},
