+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{buerkner2019,
-  title={On the Statistical and Practical Limitations of Thurstonian {IRT} Models},
+  title={On the Statistical and Practical Limitations of {T}hurstonian {IRT} Models},
   author={B{\"u}rkner, Paul-Christian and Schulte, Niklas and Holling, Heinz},
   journal={Educational and Psychological Measurement},
   year={2019},
