### 2019-10-19 #1662 @paul-buerkner
Thanks @cMadan! The bib file should be correct now.
