### 2019-10-22 #1662 @paul-buerkner
Thanks!

Tagged release is 0.10.0 and can be found at https://github.com/paul-buerkner/thurstonianIRT/releases/tag/v0.10.0

Zenodo link is: https://github.com/paul-buerkner/thurstonianIRT/releases/tag/v0.10.0

Zenodo DOI is 10.5281/zenodo.3515749
