### 2019-10-23 #1662 @cMadan
@openjournals/joss-eics, I think we're all set to accept here!
