### 2019-10-23 #1662 @kthyng
Congratulations to @paul-buerkner on your JOSS submission!!

Thank you to @tomfaulkenberry and @russellpierce for your time and expertise as reviewers, and thanks to @cMadan for editing.
