### 2019-10-23 #1662 @paul-buerkner
Thank you so much for your thorough reviewing and editing!!!

whedon <notifications@github.com> schrieb am Mi., 23. Okt. 2019, 19:25:

> 🎉🎉🎉 Congratulations on your paper acceptance! 🎉🎉🎉
>
> If you would like to include a link to your paper from your README use the
> following code snippets:
>
> Markdown:
>
>
>
>
> HTML:
>
> <a style="border-width:0" href="https://doi.org/10.21105/joss.01662">
>
>
> </a>
>
>
>
> reStructuredText:
>
>
>    :target: https://doi.org/10.21105/joss.01662
>
> This is how it will look in your documentation:
>
> [image: DOI] <https://doi.org/10.21105/joss.01662>
>
> *We need your help!*
>
> Journal of Open Source Software is a community-run journal and relies upon
> volunteer effort. If you'd like to support us please consider doing either
> one (or both) of the the following:
>
>    - Volunteering to review for us sometime in the future. You can add
>    your name to the reviewer list here:
>    https://joss.theoj.org/reviewer-signup.html
>    - Making a small donation to support our running costs here:
>    https://numfocus.org/donate-to-joss
>
